backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.0
aiosqlite==0.20.0
httpx==0.27.0
google-generativeai>=0.8.3
anthropic>=0.28.0
nanoid==2.0.0
python-dotenv==1.0.1
tenacity>=8.3.0
